requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic>=2.7,<3
jsonschema==4.21.1
jinja2==3.1.4
python-dotenv==1.0.1
openai>=1.30.0
manim==0.19.0
